Task 2

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,7 +10,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int mask;
+	unsigned long int mask;
 
 	mask = n >> index;
